requirements.txt
asgiref==3.7.2
Django==4.2.5
django-debug-toolbar==4.2.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
